Specify install instructions
Specify install instructions for the root source and lib source.

Note here the CMAKE_INSTALL_PREFIX is not defined, so in windows
it is set to "C:/Program Files(x86)/${PROJECT_NAME}" by default,
in Linux it is set to "/usr/local" by default.

That means if in CMakeLists.txt, install intruction is set as:

install(TARGETS <bin_name> DESTINATION <to_path>)

then in windows it will be installed to
"C:/Program Files(x86)/${PROJECT_NAME}/<to_path>" by default, in
Linux it will be installed to /user/local/<to_path> by default.