The 1st version modification for Step9
This modification is the 1st version of Step9 of CMake tutorial,
while it doens't have the result in the tutorial said, even this
can be compiled without errors (though a warning is issued).

Not sure what happened and what's wrong, need to think about it
later and find out why.